# swagger-spec.json
{"openapi":"3.0.0","paths":{"/running":{"get":{"operationId":"AppController_getVersionEndpoint","summary":"Verique version BFF","parameters":[],"responses":{"200":{"description":"Endpoint executed with sucess"},"400":{"description":"Invalid executed Endpoint","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpExceptionDto"}}}}}}},"/health":{"get":{"operationId":"HealthController_check","summary":"Validate BFF status","parameters":[],"responses":{"200":{"description":"Endpoint executed with sucess"},"400":{"description":"Invalid executed Endpoint","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HttpExceptionDto"}}}},"503":{"description":"The Health Check is not successful","content":{"application/json":{"schema":{"type":"object","properties":{"status":{"type":"string","example":"error"},"info":{"type":"object","example":{"database":{"status":"up"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"error":{"type":"object","example":{"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}},"nullable":true},"details":{"type":"object","example":{"database":{"status":"up"},"redis":{"status":"down","message":"Could not connect"}},"additionalProperties":{"type":"object","properties":{"status":{"type":"string"}},"additionalProperties":{"type":"string"}}}}}}}}}}}},"info":{"title":"archref-backend","description":"<< Descrição da API>>","version":"0.0.1","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"authorization":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"HttpExceptionDto":{"type":"object","properties":{"statusCode":{"type":"number","description":"Status Code","example":500},"timestamp":{"type":"string","description":"Timestamp","example":"2022-05-10T13:12:52.817Z"},"path":{"type":"string","description":"Requested Path","example":"/health"}},"required":["statusCode","timestamp","path"]}}}}